+++ lib/ui/events2/Event2SetupSurveyPanel.dart
@@ -16,6 +16,7 @@
 
 import 'dart:io';
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/ext/Event2.dart';
 import 'package:illinois/ext/Survey.dart';
@@ -31,16 +32,30 @@ import 'package:rokwire_plugin/service/events2.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/service/surveys.dart';
+import 'package:rokwire_plugin/ui/panels/survey_panel.dart';
 import 'package:rokwire_plugin/ui/popups/popup_message.dart';
-import 'package:rokwire_plugin/ui/widgets/survey.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class Event2SetupSurveyPanel extends StatefulWidget {
+  static final String routeName = 'Event2SetupSurveyPanel';
+
   final Event2SetupSurveyParam surveyParam;
   final List<Survey>? surveysCache;
   final String? eventName;
 
   Event2SetupSurveyPanel({Key? key, required this.surveyParam, this.eventName, this.surveysCache}) : super(key: key);
+
+  static Future<Event2SetupSurveyParam?> push(BuildContext context, {Key? key, required Event2SetupSurveyParam surveyParam, List<Survey>? surveysCache, String? eventName}) =>
+    Navigator.push<Event2SetupSurveyParam>(context, CupertinoPageRoute(
+      settings: RouteSettings(name: routeName),
+      builder: (context) => Event2SetupSurveyPanel(
+        surveyParam: surveyParam,
+        eventName: eventName,
+        surveysCache: surveysCache,
+      ),
+    ));
+
+  static void popUntil(BuildContext context) => Navigator.of(context).popUntil((route) => route.settings.name == routeName);
 
   Event2SurveyDetails? get details => (surveyParam.event?.id != null) ? surveyParam.event?.surveyDetails : surveyParam.details;
 
@@ -62,8 +77,6 @@ class _Event2SetupSurveyPanelState extends State<Event2SetupSurveyPanel>  {
   bool _modified = false;
   bool _loadingSurveys = false;
   bool _updatingSurvey = false;
-
-  late final SurveyWidgetController _surveyController = SurveyWidgetController();
 
   @override
   void initState() {
@@ -274,18 +287,14 @@ class _Event2SetupSurveyPanelState extends State<Event2SetupSurveyPanel>  {
 
   // Survey Preview
 
-  Widget _buildSurveyPreviewSection() => Visibility(visible: _survey != null, child:
-    Column(crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Padding(padding: EdgeInsets.only(bottom: 16), child: Container(height: 1, color: Styles().colors.fillColorPrimaryTransparent015,),),
-        Text(Localization().getStringEx('panel.event2.setup.survey.preview.title', 'Survey Preview'), style: Styles().textStyles.getTextStyle('widget.title.large.fat')),
-        Padding(
-          padding: const EdgeInsets.only(top: 8),
-          child: Event2CreatePanel.buildSectionTitleWidget(Localization().getStringEx('panel.event2.setup.survey.preview.subtitle', 'Optional: Try out the survey by filling out the sample below.')),
-        ),
-        SurveyWidget(survey: _displaySurvey, controller: _surveyController, summarizeResultRules: true, summarizeResultRulesWidget: _buildPreviewContinueWidget()),
-      ],
-    )
+  Widget _buildSurveyPreviewSection() => Visibility(visible: _survey != null,
+    child: Event2CreatePanel.buildButtonSectionWidget(
+      heading: Event2CreatePanel.buildButtonSectionHeadingWidget(
+        title: Localization().getStringEx('panel.event2.setup.survey.preview.title', 'Survey Preview'),
+        subTitle: Localization().getStringEx('panel.event2.setup.survey.preview.subtitle', 'Try out the survey by filling out a sample.'),
+        onTap: _onSurveyPreview,
+      ),
+    ),
   );
 
   Widget _buildPreviewContinueWidget() {
@@ -301,10 +310,16 @@ class _Event2SetupSurveyPanelState extends State<Event2SetupSurveyPanel>  {
       ),
       buttonTitle: Localization().getStringEx("dialog.ok.title", "OK"),
       onTapButton: (context) {
-        Navigator.pop(context);
+        Event2SetupSurveyPanel.popUntil(context);
       },
     );
   }
+
+  String get surveyTitleMacro => '{{survey_title}}';
+
+  PreferredSizeWidget? _buildPreviewHeaderBar() => HeaderBar(
+    title: Localization().getStringEx('panel.event2.setup.survey.preview.headerbar.title', '$surveyTitleMacro Preview').replaceAll(surveyTitleMacro, _displaySurvey?.title ?? Localization().getStringEx("panel.event2.setup.survey.survey.title", "SURVEY")),
+  );
 
   // HeaderBar
 
@@ -462,6 +477,15 @@ class _Event2SetupSurveyPanelState extends State<Event2SetupSurveyPanel>  {
         Navigator.of(context).pop(surveyParam);
       }
     }
+  }
+
+  void _onSurveyPreview() {
+    Navigator.of(context).push(CupertinoPageRoute(builder: (context) => SurveyPanel(
+        survey: _displaySurvey,
+        summarizeResultRules: true,
+        summarizeResultRulesWidget: _buildPreviewContinueWidget(),
+        headerBar: _buildPreviewHeaderBar(),
+    )));
   }
 }